separate handlers and server from main
